- work on data layer

+++ core/storage/src/commonMain/kotlin/data/database/entities/CheckInEntity.kt
@@ -1,13 +1,13 @@
 package data.database.entities
 
 data class CheckInEntity(
-    val id: Int = 0,
-    val emotionId: Int,
-    val factorId: Int,
-    val exercisesId: Int = 0,
+    val id: Long = 0,
+    val emotionId: Long,
+    val factorId: Long,
+    val exercisesId: Long = 0,
     val note: String = "",
     val createdAt: String,
     val createdAtLong: Long,
     val editedAt: String = "",
-    val isSynchronized: Int = 0
+    val isSynchronized: Long = 0
 )
